Bugsplat. Added exception handling for all async methods in the code, fixed all bugs found, currently no exceptions are thrown. Care should be made to keep following the guidline of keeping all async methods self contained and push exceptions into the appropriate display medium.
Known bug:
Sometimes pressing ready in the lobby after creating a game just processes then puts the button back to ready. Leaving the lobby and entering it again through the game search page allows you to ready up correctly.

+++ PhoneTag.WebServices/Controllers/ExpirationControllers/RoomExpirationController.cs
@@ -2,6 +2,8 @@
 using MongoDB.Driver;
 using PhoneTag.SharedCodebase.Events.OpLogEvents;
 using PhoneTag.SharedCodebase.Models;
+using PhoneTag.WebServices;
+using PhoneTag.WebServices.Utilities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -33,17 +35,30 @@ namespace PhoneTag.SharedCodebase.Controllers.ExpirationControllers
         //A room that already started playing doesn't expire.
         private static async Task handleRoomExpiration(ObjectId i_ExpiredId)
         {
-            GameRoom room = await RoomController.GetRoomModel(i_ExpiredId.ToString());
+            if(i_ExpiredId != null)
+            { 
+                GameRoom room = await RoomController.GetRoomModel(i_ExpiredId.ToString());
 
-            if (!room.Started)
-            {
-                foreach(String userId in room.LivingUsers)
+                if (room != null)
                 {
-                    await UsersController.LeaveRoom(userId);
-                }
+                    if (!room.Started)
+                    {
+                        foreach (String userId in room.LivingUsers)
+                        {
+                            await UsersController.LeaveRoom(userId);
+                        }
 
-                FilterDefinition<BsonDocument> roomFilter = Builders<BsonDocument>.Filter.Eq("_id", i_ExpiredId);
-                await Mongo.Database.GetCollection<BsonDocument>("Rooms").DeleteOneAsync(roomFilter);
+                        try
+                        {
+                            FilterDefinition<BsonDocument> roomFilter = Builders<BsonDocument>.Filter.Eq("_id", i_ExpiredId);
+                            await Mongo.Database.GetCollection<BsonDocument>("Rooms").DeleteOneAsync(roomFilter);
+                        }
+                        catch (Exception e)
+                        {
+                            ErrorLogger.Log(e.Message);
+                        }
+                    }
+                }
             }
         }
     }
